MMTEngine.set(): verify that config_option_exists
