+++ cli/engine.py
@@ -288,7 +288,30 @@ class MMTEngine(object):
             self._config.read(self._config_file)
         return self._config
 
+    def set(self, section, option, value=None):
+        """
+        Set engine configuration option in the config dictionary.
+        * use dependency.Injector with read_config() and inject() to affect MoseeFeatures, so an up-to-date 'moses.ini' gets written to disk in write_configs()
+        * call write_configs() to write 'engine.ini' (and 'moses.ini') to disk
+        * call ClusterNode.restart() for values to take effect
+        """
+        assert(MMTEngine.config_option_exists(section, option))
+        # coerce all types to str -- because they are parsed back in "ConfigParser.py", line 663, in _interpolate
+        self.config.set(section, option, str(value))
+
+    @staticmethod
+    def config_option_exists(section, option):
+        """check if section and option indeed exist"""
+        from cli import dependency  # cannot be at the top to avoid circular imports
+        for clazz in dependency.injectable_components:
+            if not hasattr(clazz, 'injectable_fields') or not hasattr(clazz, 'injector_section'):
+                continue
+            if clazz.injector_section == section and option in clazz.injectable_fields:
+                return True
+        return False
+
     def write_configs(self):
+        """write engine.ini and moses.ini"""
         self.moses.create_ini()
         self.write_engine_config()
